Setup/requirements.txt
# Core LangChain (modern 1.x APIs)
langchain==1.1.3
langchain-core==0.1.45
langchain-community==0.0.38

# LLM Provider (Google Gemini)
langchain-google-genai==1.0.3

# Tools
tavily-python==0.3.3

# Environment management
python-dotenv==1.0.1

# Optional but recommended (debugging / typing)
pydantic>=2.7,<3.0
typing-extensions>=4.7
